requirements.txt
pytest
requests
pyhamcrest
selenium
allure-pytest
chromedriver_py
SQLAlchemy
mysql-connector-python-rf
pexpect
paramiko